+++ config/plugin.php
@@ -13,7 +13,7 @@ function erp_define_if_not($name, $val) {
 }
 
 /* Debugging flag */
-erp_define_if_not('ERP_DEBUG', false);
+erp_define_if_not('ERP_DEBUG', true);
 
 /* Quickbooks sanbox mode flag */
 erp_define_if_not('QB_SANDBOX_MODE', ERP_DEBUG);