+++ js/app/job_request/EditJobRequestCtrl.js
@@ -101,12 +101,7 @@ function EditJobRequestCtrl(
 
     $scope.submitForm = function() {
         // Check address files changed to re-geolocation
-        if ($scope.referral.address && $scope.referral.address.length && (
-                $scope.referralForm.address.$dirty ||
-                $scope.referralForm.city.$dirty ||
-                $scope.referralForm.state.$dirty ||
-                $scope.referralForm.zip_code.$dirty)
-            ) {
+        if ($scope.referral.address && $scope.referral.address.length) {
             erpGeoLocation.resolve(getFullAddress())
                 .then(
                     function(result) {